great catch by Ryan

+++ src/lib/ui/CommunityDelete.svelte
@@ -31,7 +31,7 @@
 	};
 
 	const onKeydown = async (e: KeyboardEvent) => {
-		if (e.key === 'Enter') {
+		if (!locked && e.key === 'Enter') {
 			e.preventDefault();
 			await deleteCommunity();
 		}
